- Umsetzung der letzten verbliebenen Anforderungen. - Should be done now.

+++ app/src/main/java/de/zitruism/thl_todo_liste/interfaces/ITodoStateListener.java
@@ -2,7 +2,7 @@ package de.zitruism.thl_todo_liste.interfaces;
 
 public interface ITodoStateListener {
 
-    void updateDone(Integer id, boolean isDone);
-    void updateFavorite(Integer id, boolean isFavorite);
+    void updateDone(Long id, boolean isDone);
+    void updateFavorite(Long id, boolean isFavorite);
 
 }
